requirements.txt
# EternalMC Multi Bot Dependencies
discord.py==2.4.0
aiohttp>=3.9.0
async-timeout>=4.0.3
psutil>=6.0.0
requests>=2.31.0
pytz>=2023.3
aiosqlite>=0.19.0
